WIP - Implement shallow rendering
Preact does not directly support shallow rendering. Implement it in the
adapter by adding a VNode-hook which replaces the real child components
of a render with stubs which render a `<shallow-render>` dummy DOM element.
This dummy DOM element and the stub are then hidden from Enzyme when
converting the render result to a React Standard Tree (RST) node.

Enzyme includes logic to run lifecycle methods "manually" during shallow
rendering. This isn't needed with the approach taken to shallow
rendering here, so currently shallow rendering only works if the user
enables the `disableLifecycleMethods` option. Despite enabling that
option, lifecycle methods run as normal anyway.