Make `setState` calls apply synchronously on fully-rendered components
The `wrapper.simulate` example [1] in the Enzyme docs did not work with
Preact because the internal call to `setState` within the component
applies its changes asynchronously by default.

For consistency with how Enzyme works with React and convenience for
test authors, make `setState` calls trigger an immediate re-render when
using full rendering.

This monkey-patching currently only applies to the root component.
`setState` is still asynchronous on any stateful child components that
are rendered.

[1] https://airbnb.io/enzyme/docs/api/ReactWrapper/simulate.html