Support Preact 10 (aka. "Preact X", "ceviche")
Preact 10 is a new implementation of Preact with re-written internals.
It differs from Preact "classic" (<= 8) in several ways:

 1. VNodes more closely match React in the names of properties.
 2. Rendering diffs vnodes against the vnodes from the previous render,
    rather than the actual DOM. The "root" of the rendered-VNode tree is
    stored as a `_prevVNode` property on the container element.

    During rendering, various metadata and references to component
    instances are stored into private fields of the vnodes themselves.

    Additionally metadata about a render is stored as private properties
    on the VNode itself.
 3. The `render` function no longer takes a third argument and no longer
    returns the root element.

(2) means that a different method is needed to traverse the tree of
rendered components. Instead of traversing the DOM, the adapter instead
needs to traverse the post-render VNode tree, starting with the fragment
vnode attached to the container element passed to `render`.

This commit adds a new set of RST node-producing functions in
`src/preact10-rst.ts` for Preact 10 and modifies `MountRenderer` to
detect the version of Preact in use and use the appropriate render tree
=> RST node functions.

Additionally the render-tree => RST node conversion tests in
`MountRenderer_test.tsx` have been moved into the tests for {preact-rst,
preact10-rst}.ts, since these modules contain the actual
conversion/traversal logic.