Support fragments
Add support for fragments. The handling of fragments differs between full and
shallow rendering in the React adapters. In "mount"/full renders fragments do
not appear in the RST tree. In shallow renders they do.

This implementation follows the pattern of the full renderers in omitting
fragments from the returned RST tree.

Given a structure like:

```
<ul>
  <Fragment>
    <li/>
    <Fragment>
      <li/>
    </Fragment>
  </Fragment>
</ul>
```

The returned RST tree will look like:

```
<ul>
  <li/>
  <li/>
</ul>
```

The current implementation has a limitation that the root element passed to the
render functions must produce a single RST node. This means that it must either
not be a fragment, or must be a fragment that produces only one child.
Supporting multi-child fragments requires additional work to support the result
of `getNode` possibly being an array everywhere.