Support simulating errors in full and shallow rendering
In order to make error simulation behave realistically, the
implementation works by temporarily patching the instance's `render`
function with a stub that throws the target exception and then forcing a
re-render. After the render, the stub is removed.

Also fix a bug in testing whether anything was rendered when calling
`MountRenderer#getNode`. It is possible in Preact 10 for a component to
produce no DOM output, so we have to check for metadata on the container
instead.