Encapsulate access to internal properties of Preact types
Only access internal properties of Preact types (`__[a-z]`) in
preact-internals.ts (for Preact 10) and preact8-internals.ts (for Preact
8).

Since these internal properties are not part of the public API they are
vulnerable to being changed in new non-major releases. Ideally the adapter would
not rely on them at all. As a step towards removing this usage,
encapsulate access to these properties in specific modules and clearly
separate Preact 8 and Preact 10 internals.

Also:

 - Replace a few references to "Preact 9" with "Preact 8" since there
   was no v9 release.
 - Move `VNodeExtensions` interface out of preact-internals.ts since it
   defines properties added by the adapter, not Preact.