Rename functions for accessing render output
Split the function for accessing VNode rendered into a container and
getting the last vnode rendered by a component into two separate
functions. Although these happen to use the same code, they are logically
different operations. This also provides an opportunity to improve the
names to clarify how they are used.

Also remove access to the `__v` property of Component instances which
was not necessary as the code that called this function already had that
VNode available.