Flush state updates and effects after simulated events
Discussions on the Enzyme repo suggest that `simulate(...)` is a
semi-deprecated API, but it is the obvious way to dispatch an event
according to the docs and still the most convenient way to do so.

Flushing state updates and effects after a simulated event dispatch
avoids confusion for users about why their component hasn't updated.