Fix exception when determining type of arrow function component
Fix an exception when determining the type of a component if the
component was declared with an arrow function. Native (untranspiled)
arrow functions do not have a `prototype` property.

The existing tests would have encountered the issue, except that they
used the same transpilation settings (to ES5) as for distribution.

Fixes #37