Update Preact 10 to 10.0.0-beta.2
The representation of text nodes changed in this release due to
https://github.com/preactjs/preact/pull/1600.

The `text` property is no longer present on vnodes. Instead strings or
numbers passed to `createElement` get converted to vnodes with type
`null` and `vnode.props` equal to the string or number value.

This PR is backwards-compatible with earlier betas.

Fixes #48