Make shallow rendering compatible with Enzyme 3.10
The shallow renderer in Enzyme never took a callback argument, that was
a mistake in this adapter. In Enzyme 3.10 however a third argument has
been added which contains additional rendering options, currently
limited to a `providerValues` map. This caused an error because the code
assumed this argument was a callback and threw an error if it was not a
callback.

This commit makes the minimal changes needed to make the existing tests
pass with Enzyme 3.10. Supporting shallow rendering with new-style
context will be done separately.