Adapt tests to work with Preact v10 JSX typings
In the Preact v10 types, the JSX namespace is contained within the
preact namespace, so needs to be referenced as `preact.JSX` rather than
`JSX`. The TypeScript compiler determines which namespace to look up the
`JSX` namespace in based on the `jsxFactory` config. It first tries the
namespace containing the `jsxFactory` function and then falls back to
the global namespace.

Therefore, we need to use `preact.h` rather than just `h` as the
`jsxFactory` in order for the compiler to use `preact.JSX` to look up
types for native HTML elements.

A couple of tests that used function components without an explicit type
also had to be modified to compile against the new typings.