Use local version of package to test examples
Use the version of the adapter package from the current git checkout to
run the examples, instead of a version installed from npm.

This means that CI builds will verify that the version of the package in
the current tree works with the example, not just the version published
to npm.

It also avoids a problem when examples depend on changes in the adapter
package that have not yet been published to npm.