Add ES module build
Change the main build to generate ES modules rather than CommonJS. The
CommonJS build is still generated for backwards compatibility and will
be removed in a future major release.

The tests currently use the CommonJS build because that was easier to
get working. They should be changed to use the ESM build in future.