Upgrade Preact dependency used in tests to v10.5.15
This fixes an error I observed locally in one test:

```
  1) preact10-rst
       getNode
         returns expected RST node (component that renders children):
     TypeError: Found non-callable @@iterator
```

I didn't dig into the exact cause of the above, but the test did not fail when
run recently in CI and I have upgraded to Node v17 locally since. Using
a current version of Preact in tests is a good idea in any case.

In the process another test started failing due to the addition of an
internal ID counter to VNodes. This test was adapted to ignore the field
when comparing values.